# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Akademiyati learning platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /student/feed:
    get:
      operationId: getStudentFeed
      tags: [student]
      summary: Get a personalized short video feed
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Short video feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShortVideo"
  /teachers:
    get:
      operationId: listTeachers
      tags: [teachers]
      summary: Search teacher academies
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: subject
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Teacher academies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teacher"
    post:
      operationId: createTeacher
      tags: [teachers]
      summary: Create a teacher profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherInput"
      responses:
        "201":
          description: Created teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
  /teachers/{teacherId}:
    get:
      operationId: getTeacher
      tags: [teachers]
      summary: Get a teacher academy
      parameters:
        - $ref: "#/components/parameters/TeacherId"
      responses:
        "200":
          description: Teacher academy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
  /teachers/{teacherId}/dashboard:
    get:
      operationId: getTeacherDashboard
      tags: [teachers]
      summary: Get teacher analytics
      parameters:
        - $ref: "#/components/parameters/TeacherId"
      responses:
        "200":
          description: Teacher dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherDashboard"
...
  /teachers/{teacherId}/students:
    get:
      operationId: listTeacherStudents
      tags: [teacher-management]
      summary: List students enrolled in a teacher academy
      parameters:
        - $ref: "#/components/parameters/TeacherId"
      responses:
        "200":
          description: Teacher students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeacherStudent"
  /teachers/{teacherId}/students/{studentId}:
    get:
      operationId: getTeacherStudent
      tags: [teacher-management]
      summary: Get a student's progress in a teacher academy
      parameters:
        - $ref: "#/components/parameters/TeacherId"
        - name: studentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Student progress detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherStudentDetail"
  /teachers/{teacherId}/revenue:
    get:
      operationId: getTeacherRevenue
      tags: [teacher-management]
      summary: Get demo teacher revenue
      parameters:
        - $ref: "#/components/parameters/TeacherId"
      responses:
        "200":
          description: Revenue summary and transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherRevenue"
  /teachers/{teacherId}/content:
    get:
      operationId: listTeacherContent
      tags: [content]
      summary: List teacher content
      parameters:
        - $ref: "#/components/parameters/TeacherId"
      responses:
        "200":
          description: Content list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentItem"
    post:
      operationId: createContent
      tags: [content]
      summary: Publish teacher content
      parameters:
        - $ref: "#/components/parameters/TeacherId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentInput"
      responses:
        "201":
          description: Published content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentItem"
  /teachers/{teacherId}/content/{contentId}:
    patch:
      operationId: updateTeacherContent
      tags: [content, teacher-management]
      summary: Update teacher content
      parameters:
        - $ref: "#/components/parameters/TeacherId"
        - name: contentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentUpdate"
      responses:
        "200":
          description: Updated content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentItem"
    delete:
      operationId: deleteTeacherContent
      tags: [content, teacher-management]
      summary: Delete teacher content
      parameters:
        - $ref: "#/components/parameters/TeacherId"
        - name: contentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Content deleted
  /teachers/{teacherId}/short-videos:
    post:
      operationId: createTeacherShortVideo
      tags: [content, teacher-management]
      summary: Publish a short video to the student feed
      parameters:
        - $ref: "#/components/parameters/TeacherId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShortVideoInput"
      responses:
        "201":
          description: Published short video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShortVideo"
  /academies/{teacherId}/enroll:
    post:
      operationId: enrollInAcademy
      tags: [teachers]
      summary: Enroll the current student in an academy
      parameters:
        - $ref: "#/components/parameters/TeacherId"
      responses:
        "200":
          description: Enrollment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
  /quizzes:
    get:
      operationId: listQuizzes
      tags: [quizzes]
      summary: List available quizzes
      responses:
        "200":
          description: Quiz list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuizSummary"
    post:
      operationId: createQuiz
      tags: [quizzes]
      summary: Create a quiz
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizInput"
      responses:
        "201":
          description: Created quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quizzes/{quizId}:
    get:
      operationId: getQuiz
      tags: [quizzes]
      summary: Get a quiz with questions
      parameters:
        - $ref: "#/components/parameters/QuizId"
      responses:
        "200":
          description: Quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quizzes/{quizId}/attempts:
    post:
      operationId: submitQuizAttempt
      tags: [quizzes]
      summary: Submit quiz answers
      parameters:
        - $ref: "#/components/parameters/QuizId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizAttemptInput"
      responses:
        "200":
          description: Graded attempt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizAttempt"
  /students/{studentId}/progress:
    get:
      operationId: getStudentProgress
      tags: [student]
      summary: Get student learning progress
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Student progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProgress"
  /community/{teacherId}/posts:
    get:
      operationId: listCommunityPosts
      tags: [community]
      summary: List community posts
      parameters:
        - $ref: "#/components/parameters/TeacherId"
      responses:
        "200":
          description: Community posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createCommunityPost
      tags: [community]
      summary: Create a community post
      parameters:
        - $ref: "#/components/parameters/TeacherId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
...
  /admin/subscriptions/{studentId}:
    patch:
      operationId: updateStudentSubscription
      tags: [admin]
      summary: Change a student's demo subscription
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionUpdate"
      responses:
        "200":
          description: Updated subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
components:
  parameters:
    TeacherId:
      name: teacherId
      in: path
      required: true
      schema:
        type: string
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50
  schemas:
...
    Teacher:
      type: object
      required: [id, name, subject, grades, bio, studentsCount, rating, avatar, verified]
      properties:
        id: { type: string }
        name: { type: string }
        subject: { type: string }
        grades:
          type: array
          items: { type: string }
        bio: { type: string }
        studentsCount: { type: integer }
        rating: { type: number }
        avatar: { type: string }
        verified: { type: boolean }
    TeacherInput:
      type: object
      required: [name, subject, bio]
      properties:
        name: { type: string }
        subject: { type: string }
        bio: { type: string }
...
    ShortVideo:
      type: object
      required: [id, title, teacher, subject, topic, thumbnail, duration, views, likes, saved]
      properties:
        id: { type: string }
        title: { type: string }
        teacher: { $ref: "#/components/schemas/Teacher" }
        subject: { type: string }
        topic: { type: string }
        thumbnail: { type: string }
        duration: { type: string }
        views: { type: integer }
        likes: { type: integer }
        saved: { type: boolean }
...
    ContentItem:
      type: object
      required: [id, title, type, subject, grade, topic, publishedAt, views]
      properties:
        id: { type: string }
        title: { type: string }
        type: { $ref: "#/components/schemas/ContentType" }
        subject: { type: string }
        grade: { type: string }
        topic: { type: string }
        publishedAt: { type: string }
        views: { type: integer }
        duration: { type: ["string", "null"] }
        thumbnail: { type: ["string", "null"] }
...
    ContentUpdate:
      type: object
      properties:
        title: { type: string }
        type: { $ref: "#/components/schemas/ContentType" }
        subject: { type: string }
        grade: { type: string }
        topic: { type: string }
        duration: { type: string }
...
    TeacherDashboard:
      type: object
      required: [studentsCount, views, subscribersCount, avgLessonWatch, avgQuizScore, activeStudents, averageScore, revenue, topContent, topMistakes]
      properties:
        studentsCount: { type: integer }
        views: { type: integer }
        subscribersCount: { type: integer }
        avgLessonWatch: { type: integer }
        avgQuizScore: { type: integer }
        activeStudents: { type: integer }
        averageScore: { type: integer }
        revenue: { type: number }
        topContent:
          type: array
          items: { $ref: "#/components/schemas/ContentItem" }
        topMistakes:
          type: array
          items:
            $ref: "#/components/schemas/MistakeTopic"
...
    TeacherStudent:
      type: object
      required: [id, name, level, lastActivity, averageScore, completion]
      properties:
        id: { type: string }
        name: { type: string }
        level: { type: string }
        lastActivity: { type: string }
        averageScore: { type: integer }
        completion: { type: integer }
    TeacherStudentDetail:
      allOf:
        - $ref: "#/components/schemas/TeacherStudent"
        - type: object
          required: [subjects, recentAttempts, completedLessons]
          properties:
            subjects:
              type: array
              items: { $ref: "#/components/schemas/ProgressSubject" }
            recentAttempts:
              type: array
              items: { $ref: "#/components/schemas/QuizAttempt" }
            completedLessons: { type: integer }
...
    TeacherRevenue:
      type: object
      required: [totalSales, netRevenue, subscriptions, activeSubscriptions, transactions]
      properties:
        totalSales: { type: number }
        netRevenue: { type: number }
        subscriptions: { type: integer }
        activeSubscriptions: { type: integer }
        transactions:
          type: array
          items: { $ref: "#/components/schemas/RevenueTransaction" }
...